+++ engine/thread/Semaphore.hpp
@@ -36,7 +36,7 @@ namespace ouzel::thread
         {
 #ifdef _MSC_VER
             if (semaphore == INVALID_HANDLE_VALUE)
-                throw std::system_error{GetLastError(), std::system_category(), "Failed to create semaphore"};
+                throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to create semaphore"};
 #elif defined(__APPLE__)
             if (!semaphore)
                 throw std::runtime_error{"Failed to create semaphore"};
@@ -67,7 +67,7 @@ namespace ouzel::thread
         {
 #ifdef _MSC_VER
             if (WaitForSingleObject(semaphore, INFINITE) == WAIT_FAILED)
-                throw std::system_error{GetLastError(), std::system_category(), "Failed to wait for semaphore"};
+                throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to wait for semaphore"};
 #elif defined(__APPLE__)
             if (dispatch_semaphore_wait(semaphore, DISPATCH_TIME_FOREVER) != 0)
                 throw std::runtime_error{"Failed to wait for semaphore"};
@@ -82,7 +82,7 @@ namespace ouzel::thread
         {
 #ifdef _MSC_VER
             if (!ReleaseSemaphore(semaphore, static_cast<LONG>(count), nullptr))
-                throw std::system_error{GetLastError(), std::system_category(), "Failed to release semaphore"};
+                throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to release semaphore"};
 #elif defined(__APPLE__)
             while (count-- > 0)
                 if (dispatch_semaphore_signal(semaphore) != 0)
